requirements.txt
Django >=1.9,<1.10
# django-crispy-forms >=1.6.0
-e git://github.com/maraujop/django-crispy-forms.git#egg=dev
django-static-precompiler >=1.4
django-model-utils >= 2.5
django-netfields >=0.5
django-colorfield >=0.1.10
djangorestframework >= 3.3.0
django-celery-email >=1.1

celery >=3.1
py-pgp
